+++ app/src/main/kotlin/com/masterofchessstrategy/game/ChineseChessGameViewModel.kt
@@ -97,7 +97,6 @@ class ChineseChessGameViewModel internal constructor(
 
     override fun onCleared() {
         closeEngine()
-        super.onCleared()
     }
 
     private fun selectPiece(